prefs/appearance: move background opacity settings
Move background opacity settings from Terminal/Colors to
Window/Appearance, because it controls whole window's background
opacity, not just the terminal area.
